requirements-dataset.txt
# Dependencies for Module 1 (dataset preparation) and visualization.
opencv-python>=4.8.0
PyYAML>=6.0
numpy>=1.24.0
